--- internal/api/corporate_actions.go
@@ -1,1 +1,0 @@
-package api
