--- internal/api/order.go
@@ -1,1 +1,0 @@
-package api
