--- internal/api/position.go
@@ -1,1 +1,0 @@
-package api
