--- internal/api/watchlist.go
@@ -1,1 +1,0 @@
-package api
