+++ internal/config/config.go
@@ -5,7 +5,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"os"
-	"strings"
 
 	kitlog "github.com/go-kit/log"
 
@@ -24,10 +23,11 @@ type GoogleCloud struct {
 }
 
 type Alpaca struct {
-	APIKey       string `json:"api_key" omitempty:"true"`
-	APISecretKey string `json:"api_secret_key" omitempty:"true"`
-	APIBaseURL   string `json:"api_base_url" omitempty:"true"`
-	APIDataURL   string `json:"api_data_url" omitempty:"true"`
+	APIKey         string `json:"api_key" omitempty:"true"`
+	APISecretKey   string `json:"api_secret_key" omitempty:"true"`
+	APIBaseURL     string `json:"api_base_url" omitempty:"true"`
+	APIDataURL     string `json:"api_data_url" omitempty:"true"`
+	MarketDataFeed string `json:"market_data_feed" omitempty:"true"`
 }
 
 type Crowemi struct {
@@ -38,31 +38,11 @@ type Crowemi struct {
 	DatabaseURI     string            `json:"database_uri" omitempty:"true"`
 }
 
-type Scheduler struct {
-	ScheduleTasks []SchedulerTask `json:"scheduler_tasks" omitempty:"true"`
-}
-
-type SchedulerTask struct {
-	Name     string `json:"name" omitempty:"true"`
-	Schedule string `json:"schedule" omitempty:"true"`
-}
-
-func (s Scheduler) ScheduleForTask(taskName string) string {
-	for _, task := range s.ScheduleTasks {
-		if strings.EqualFold(strings.TrimSpace(task.Name), strings.TrimSpace(taskName)) {
-			return strings.TrimSpace(task.Schedule)
-		}
-	}
-	return ""
-}
-
 type Runtime struct {
-	PortfolioID        string    `json:"portfolio_id" omitempty:"true"`
-	HTTPListenAddr     string    `json:"http_listen_addr" omitempty:"true"`
-	Scheduler          Scheduler `json:"scheduler" omitempty:"true"`
-	TaskTimeout        string    `json:"task_timeout" omitempty:"true"`
-	StreamReconnectMin string    `json:"stream_reconnect_min" omitempty:"true"`
-	StreamReconnectMax string    `json:"stream_reconnect_max" omitempty:"true"`
+	PortfolioID string    `json:"portfolio_id" omitempty:"true"`
+	Server      Server    `json:"server" omitempty:"true"`
+	Scheduler   Scheduler `json:"scheduler" omitempty:"true"`
+	Streamer    Streamer  `json:"streamer" omitempty:"true"`
 }
 
 type Notifier struct {
@@ -110,22 +90,5 @@ func Bootstrap(configPath string) (*Config, error) {
 	logger := kitlog.NewLogfmtLogger(kitlog.NewSyncWriter(os.Stderr))
 	logger = kitlog.With(logger, "ts", kitlog.DefaultTimestampUTC, "caller", kitlog.DefaultCaller)
 	config.Logger = logger
-	applyRuntimeDefaults(&config)
-
 	return &config, nil
-}
-
-func applyRuntimeDefaults(config *Config) {
-	if config.Runtime.HTTPListenAddr == "" {
-		config.Runtime.HTTPListenAddr = ":8080"
-	}
-	if config.Runtime.TaskTimeout == "" {
-		config.Runtime.TaskTimeout = "30s"
-	}
-	if config.Runtime.StreamReconnectMin == "" {
-		config.Runtime.StreamReconnectMin = "1s"
-	}
-	if config.Runtime.StreamReconnectMax == "" {
-		config.Runtime.StreamReconnectMax = "30s"
-	}
 }
