+++ internal/config/server.go
@@ -1,0 +1,5 @@
+package config
+
+type Server struct {
+	Enabled bool `json:"enabled" omitempty:"true"`
+}
