+++ internal/runtime/app.go
@@ -6,17 +6,19 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/crowemi-io/crowemi-trades/internal/scheduler"
+	"github.com/crowemi-io/crowemi-trades/internal/stream"
 	kitlog "github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"golang.org/x/sync/errgroup"
 )
 
 type App struct {
-	Logger       kitlog.Logger
-	Server       *http.Server
-	Scheduler    interface{ Run(context.Context) error }
-	Stream       interface{ Run(context.Context) error }
-	TradeUpdates interface{ Run(context.Context) error }
+	Logger    kitlog.Logger
+	Server    *http.Server
+	Scheduler *scheduler.Runner
+	Watcher   *stream.Watcher
+	Updater   *stream.Updater
 }
 
 func (a *App) Run(ctx context.Context) error {
@@ -48,17 +50,17 @@ func (a *App) Run(ctx context.Context) error {
 		})
 	}
 
-	if a.Stream != nil {
+	if a.Watcher != nil {
 		group.Go(func() error {
-			_ = level.Info(a.Logger).Log("component", "stream", "msg", "consumer start")
-			return a.Stream.Run(groupCtx)
+			_ = level.Info(a.Logger).Log("component", "watcher", "msg", "consumer start")
+			return a.Watcher.Run(groupCtx)
 		})
 	}
 
-	if a.TradeUpdates != nil {
+	if a.Updater != nil {
 		group.Go(func() error {
-			_ = level.Info(a.Logger).Log("component", "trade_updates", "msg", "consumer start")
-			return a.TradeUpdates.Run(groupCtx)
+			_ = level.Info(a.Logger).Log("component", "updater", "msg", "consumer start")
+			return a.Updater.Run(groupCtx)
 		})
 	}
